Fix benchmark script
