+++ .github/workflows/create-release-after-tag.yml
@@ -271,8 +271,7 @@ jobs:
 
       - name: Run Upload Script
         run: |
-          cd scripts
-          python3 upload_federalist.py ${{ github.ref }}
+          python3 ./scripts/upload_federalist.py ${{ github.ref }}
         env:
           AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID  }}
           AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY  }}
