+++ stork-lib/benches/basic.rs
@@ -1,15 +1,20 @@
 use criterion::{criterion_group, criterion_main, Criterion};
-use std::{convert::TryFrom, path::PathBuf, time::Duration};
+use std::{convert::TryFrom, path::PathBuf, process::exit, time::Duration};
 use stork_config::Config;
 
 fn config_from_path(path: &str) -> Config {
+    if !std::path::Path::join(&std::env::current_dir().unwrap(), ".stork-project-root").exists() {
+        println!("To successfully run this benchmark, the working directory must be the Stork project root.\nIt looks like the working directory is {:?}.\nRunning `just bench` will do this automatically.", std::env::current_dir());
+        exit(50);
+    }
+
     let path = PathBuf::from(path);
     let contents = std::fs::read_to_string(path).unwrap();
     return Config::try_from(contents.as_str()).unwrap();
 }
 
 fn build_federalist(c: &mut Criterion) {
-    let config = config_from_path("./benches/federalist.toml");
+    let config = config_from_path("./stork-lib/benches/federalist.toml");
 
     let mut group = c.benchmark_group("build");
     group.measurement_time(Duration::from_secs(12));
@@ -20,7 +25,7 @@ fn build_federalist(c: &mut Criterion) {
 }
 
 fn search_federalist_for_liberty(c: &mut Criterion) {
-    let config = config_from_path("./benches/federalist.toml");
+    let config = config_from_path("./stork-lib/benches/federalist.toml");
     let index = stork_lib::V3Build(&config).unwrap().index;
 
     let mut group = c.benchmark_group("search/federalist");
@@ -28,9 +33,9 @@ fn search_federalist_for_liberty(c: &mut Criterion) {
 
     let queries = vec![
         "liberty",
-        "lib",
-        "liber old world",
-        "some long query that won't return results but let's see how it does",
+        // "lib",
+        // "liber old world",
+        // "some long query that won't return results but let's see how it does",
     ];
 
     for query in &queries {
